More files for reducers and actions for the help page.
