+++ client/app/help/helpConstants.js
@@ -1,0 +1,4 @@
+export const ACTIONS = {
+  SET_ORGANIZATIONS: 'SET_ORGANIZATIONS',
+  SET_ORGANIZATION_MEMBERSHIP_REQUESTS: 'SET_ORGANIZATION_MEMBERSHIP_REQUESTS',
+};
