--- client/app/help/reducers.js
@@ -1,24 +1,0 @@
-import { update } from '../util/ReducerUtil';
-
-export const initialState = {
-  messages: {
-    success: null,
-    error: null
-  },
-  featureToggles: {},
-  userRole: '',
-  userCssId: '',
-  userInfo: null,
-  organizations: [],
-  activeOrganization: {
-    id: null,
-    name: null,
-    isVso: false
-  },
-  userIsVsoEmployee: false,
-  userIsCamoEmployee: false,
-  feedbackUrl: '#',
-  loadedUserId: null,
-};
-
-
